WP 4.6 Compat: set WP_HOME/SITEURL directly
https://github.com/WordPress/WordPress/commit/905f4ec0f835dc9eb0a34f37c297bd6a02d519ba
changed email address handling with PHPMailer. It's not validating email
addresses. This was failing since Trellis was using string interpolation
with the `HTTP_HOST` constant. It wasn't being evaluated in that
context.

The solution is to set these values directly with the first canonical
site host eliminating the need for the magic constant and string
interpolation.