Update MariaDB repo and key for Xenial
MariaDB was still using the repo and signing keys for Trusty which was also using weak digest algorithm (SHA 1).

This is safe to run on existing installs because it will detect if you have the Trusty apt repo currently configured, conditionally remove those lines, conditionally upgrade mariadb-client mariadb-server to the correct version but only if you had the Trusty repo previously configured so it won't upgrade those packages on every run.