Begin adding configuration for SSL termination
This PR adds the `load_balancer_enabled` and `ssl_terminated` helpers.
It checks for the `load_balancer` key on a site. For example:

```
wordpress_sites:
  example.com:
    // blah blah
    ssl:
      enabled: false
    load_balancer:
      enabled: true
      ssl_terminated: true
    // blah blah
```

The pattern it follows is the load balancer will send https traffic to
port 80 and http traffic to port 8080. Because SSL has already been
handled, the server doesn't need to do anything over 443, so it can be
treated like a normal web request (thus leaving everything alone and
communicating over port 80). Port 8080 is then used to receive http
traffic and then force redirect to the https version of the site. To
coerce WordPress into thinking this is an HTTPS request, we set the
`HTTPS` PHP `$_SERVER` variable to make `is_ssl()` work inside
WordPress.