Check Ansible version before Ansible validates task attributes
Avoids "ERROR! ‘check_mode’ is not a valid attribute for a Task"
which occurs for Ansible lt 2.2. Ansible validates task
attributes before running a playbook and before loading callback
plugins, so a vars plugin appeared to be the only option for
intercepting the error with a version check and message.