Remove db_import option
This option is used to import a local SQL dump file into the database.

It was seldom used and could be dangerous. This option would cause the
DB to be imported *every* provision which could overwrite data.

As an alternative for most DB operations, we always recommend using WP
CLI.

Example: `wp db import dbname.sql`