Remove WP transient _site_transient_theme_roots during deploys
WP options template_root and stylesheet_root will not exist for
a site when no theme has been activated explicitly, when a site
defaults to a theme like twentyseventeen. In this case, WP consults
the _site_transient_theme_roots transient for the path to the theme.
This transient contains release paths that do not update on deploy,
leading to WSOD due to various instances of "No such file or directory."
This commit avoids the problem by removing the transient immediately
after the new release path is symlinked. WP auto-regenerates the
transient with correct paths.