Allow Trellis VMs to be created/provisioned in Windows Subsystem for Linux.
WSL does not support NFS, but Trellis's checks to determine whether or not to
use NFS for Vagrant shares assume that NFS is supported if the operating system
is not reported to be Windows. Unfortunately this means that it attempts to
use NFS in WSL and consequently fails..

This change simply adds an additional check: If Vagrant believes that it is
being run from WSL, it syncs folders without NFS (just as though it was
running in Windows without `vagrant-winnfsd`).

Note: This applies to Vagrant when it is run as a Linux executable through
WSL, *not* when the Windows `vagrant.exe` is run from WSL.