Improve failed_when rule for Wordpress Installed check
In rare cases the wp_installed registered var may be missing the
stderr attribute, so add a default to avoid related error.

The `wp core  is-installed` command return code is 1 if WP is simply
not installed. However, in rare cases the command may return some
other return code indicative of true failure, so fail if rc > 1.