Remove ansible_python_interpreter in server.yml
This was changing the interpreter/version of Python used on remote
hosts in `server.yml` to support Python 2 or 3.

However, this doesn't work in all cases yet because we're installing
Python 2 specific packages on servers.

So for now Python 3 support is limited to the host/control machine which
is the bigger issue anyway. Even Ubuntu 18.04 still defaults to Python
2.