Use all canonical site hosts for Landrush TLD (#1077)
* Use all canonical site hosts for Landrush TLD

This is useful if you want to use WordPress’ multisite domain mapping
locally and if you’ve specified canonical hosts which are not
sub-domains of the main canonical host.

* Filter subdomains for Landrush TLD resolver

This filters out subdomains of the main hostname for the Landrush
TLD(s). This way it only adds necessary resolvers.

For example, of the following:

- `example.test` (main hostname)
- `subdomain.example.test`
- `client.test`

We tell Landrush (dnsmaq) to only create:

- `/etc/resolver/example.test`
- `/etc/resolver/client.test`

Because `example.test` handles `subdomain.example.test`.