| bool thing #1199
Regarding #1199 but there may be other places where a `list` type got a `| bool` added that needs fixing.

I guess this is the way to do it. Not an Ansible expert though. Could also add another check if the item is actually not empty and a string or something but ...