Using variable for location instead of hard-coded
Currently `web` is hard-coded location path for nginx. This is true for bedrock. For sites which are not bedrock specific, it will be helpful to keep it configurable.
I know trellis only supports wordpress, but I am using this for even setting up laravel projects where the location path is `public` instead of `web`. I manually have to modify the nginx config, whenever I run ansible playbook for server.yml.