Remove explicit permission for site directory
Fixes #1311

The `mode` option was set in https://github.com/roots/trellis/pull/1270
for linting purposes. However, this can result in the permissions
changing on local site files causing them to appear as changed in Git.

Since this directory is guaranteed to exist, we don't need to set
`mode`. The purpose of this task is just to set the owner + group
recursively.